Updated seed file so Vraj can work on the map side of things
Removed most of the seed images and added 2 photos from 5 different locations. Those 2 photos are, right now, in the same spot, but I changed the coordinates to make them approximately 200 miles away from each other. Kept the 2 photos that I took in my back yard with location services on to make sure photos that already include the metadata that we need are processed correctly.